[Refactor] #44 - Refactoring Login Button Tags
