+++ src/components/login/LoginBottomsheetContent.styled.ts
@@ -22,24 +22,3 @@ export const LoginBottomsheetContentTopSubTitle = styled.div`
   ${({ theme }) => theme.fonts.b1}
   color: ${({ theme }) => theme.colors.font.blackLight};
 `;
-
-export const LoginBottomsheetContentButton = styled.button`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  width: 100%;
-  height: 3.5rem;
-  padding: 0 1.25rem;
-  gap: 0.5rem;
-  border-radius: 0.5rem;
-
-  background-color: #fee500;
-
-  cursor: pointer;
-`;
-
-export const LoginBottomsheetContentButtonText = styled.span`
-  ${({ theme }) => theme.fonts.btn}
-  color: ${({ theme }) => theme.colors.font.black};
-`;
